QML-C++: merge ZcMathObject with MathObject, phase 1/2
Simplifies the interface and approach.

The sole purpose of the existence of ZcMathObject is so the
SyntaxHighlighter of ZcExprEdit can set the new text and get the
new state back. We achieve that now using a common virtual base
among "highlightable" backends.

Next step: merge ZcMathObject and MathObject and handle the
consequences